fatia queue_create.go em queue_create + queue_common + queue_omniq
Estrutura escalavel pra multiplos drivers de fila. Antes tudo vivia em
um arquivo de ~950 linhas mistudando dispatcher, builders agnosticos e
template omniq. Agora:

- queue_create.go (152 linhas): API publica, tipos, mapa queueDrivers
  e funcao QueueCreate dispatcher. Eh o unico arquivo que precisa
  mudar quando um driver novo eh registrado.
- queue_common.go (206 linhas): builders driver-agnosticos
  (repositories.go, services.go que sao iguais pra qualquer driver)
  + helpers AST compartilhados (stampAssignStmt, stampCallStmt,
  prefixedConfigCall, etc.).
- queue_omniq.go (614 linhas): builders especificos do omniq
  (setup, configs, pkg, worker, main, .env). Auto-contido — adicionar
  sqs/rabbitmq/kafka eh criar queue_<driver>.go + 1 entrada no
  queueDrivers map em queue_create.go.

Bonus: supportedDriversList() formata os drivers registrados em ordem
alfabetica na mensagem de erro de driver invalido, entao a lista
cresce automaticamente conforme novos drivers entram no mapa.

Saida do gerador nao muda — testes existentes passam sem ajuste.
